requirements/testing.txt
-r common.txt

pytest
pytest-mock
mock-open
Faker